backend/requirements.txt
fastapi
uvicorn
openai
mcp
python-dotenv
pyodbc
requests
pyjwt
httpx>=0.27.0
pydantic>=2.6.0